move jibriselenium2 back to jibriselenium
