+++ src/main/kotlin/org/jitsi/jibri/selenium/JibriSelenium.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright @ 2018 Atlassian Pty Ltd
+ * Copyright @ 2018 - present 8x8, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,7 +12,6 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
 package org.jitsi.jibri.selenium
@@ -22,7 +21,7 @@ import org.jitsi.jibri.config.XmppCredentials
 import org.jitsi.jibri.selenium.pageobjects.CallPage
 import org.jitsi.jibri.selenium.pageobjects.HomePage
 import org.jitsi.jibri.selenium.util.BrowserFileHandler
-import org.jitsi.jibri.service.JibriServiceStatus
+import org.jitsi.jibri.status.ComponentState
 import org.jitsi.jibri.util.StatusPublisher
 import org.jitsi.jibri.util.TaskPools
 import org.jitsi.jibri.util.extensions.error
@@ -35,6 +34,7 @@ import org.openqa.selenium.chrome.ChromeOptions
 import org.openqa.selenium.logging.LogType
 import org.openqa.selenium.logging.LoggingPreferences
 import org.openqa.selenium.remote.CapabilityType
+import org.openqa.selenium.remote.UnreachableBrowserException
 import java.util.concurrent.ScheduledFuture
 import java.util.concurrent.TimeUnit
 import java.util.logging.Level
@@ -98,10 +98,11 @@ val RECORDING_URL_OPTIONS = listOf(
  */
 class JibriSelenium(
     private val jibriSeleniumOptions: JibriSeleniumOptions = JibriSeleniumOptions()
-) : StatusPublisher<JibriServiceStatus>() {
+) : StatusPublisher<ComponentState>() {
     private val logger = Logger.getLogger(this::class.qualifiedName)
     private var chromeDriver: ChromeDriver
     private var currCallUrl: String? = null
+    private val stateMachine = SeleniumStateMachine()
 
     /**
      * A task which executes at an interval and checks various aspects of the call to make sure things are
@@ -112,6 +113,7 @@ class JibriSelenium(
 
     companion object {
         private val browserOutputLogger = getLoggerWithHandler("browser", BrowserFileHandler())
+        const val COMPONENT_ID = "Selenium"
     }
 
     /**
@@ -147,6 +149,7 @@ class JibriSelenium(
             EmptyCallStatusCheck(),
             MediaReceivedStatusCheck(logger)
         )
+        stateMachine.onStateTransition(this::onSeleniumStateChange)
     }
 
     /**
@@ -159,22 +162,25 @@ class JibriSelenium(
         }
     }
 
+    private fun onSeleniumStateChange(oldState: ComponentState, newState: ComponentState) {
+        logger.info("Selenium transition from state $oldState to $newState")
+        publishStatus(newState)
+    }
+
     private fun startRecurringCallStatusChecks() {
         recurringCallStatusCheckTask = TaskPools.recurringTasksPool.scheduleAtFixedRate(15, TimeUnit.SECONDS, 15) {
             val callPage = CallPage(chromeDriver)
             try {
-                callStatusCheckLoop@for (callStatusCheck in callStatusChecks) {
-                    val checkResult = callStatusCheck.run(callPage)
-                    if (checkResult.shouldTerminateCall()) {
-                        recurringCallStatusCheckTask?.cancel(false)
-                        if (checkResult.isError()) {
-                            publishStatus(JibriServiceStatus.ERROR)
-                        } else {
-                            publishStatus(JibriServiceStatus.FINISHED)
-                        }
-                        // We don't want to execute any further checks
-                        break@callStatusCheckLoop
-                    }
+                // Run through each of the checks.  If we hit one that returns an event, then we stop and process the
+                // state transition from that event. Note: it's intentional that we stop at the first check that fails
+                // and 'asSequence' is necessary to do that.
+                val event = callStatusChecks
+                        .asSequence()
+                        .map { check -> check.run(callPage) }
+                        .firstOrNull { result -> result != null }
+                if (event != null) {
+                    logger.info("Recurring call status checks genereated event $event")
+                    stateMachine.transition(event)
                 }
             } catch (t: Throwable) {
                 logger.error("Error while running call status checks", t)
@@ -182,8 +188,13 @@ class JibriSelenium(
                     // We've found that a timeout here often implies chrome has hung so consider this as an error
                     // which will result in this Jibri being marked as unhealthy
                     logger.error("Javascript timed out, assuming chrome has hung")
-                    publishStatus(JibriServiceStatus.ERROR)
+                    stateMachine.transition(SeleniumEvent.ChromeHung)
                 }
+            } catch (u: UnreachableBrowserException) {
+                //TODO: i've seen this happen when we ran checks right as we were shutting down, but maybe we'd
+                // also get it if chrome crashed?  maybe we need to have some 'shuttingDown' flag we set that
+                // we could check here to know if it's an error or to just ignore it?
+                println("Check running right as we quit(?)")
             }
         }
     }
@@ -204,26 +215,30 @@ class JibriSelenium(
     /**
      * Join a a web call with Selenium
      */
-    fun joinCall(callUrlInfo: CallUrlInfo, xmppCredentials: XmppCredentials? = null): Boolean {
-        HomePage(chromeDriver).visit(callUrlInfo.baseUrl)
+    fun joinCall(callUrlInfo: CallUrlInfo, xmppCredentials: XmppCredentials? = null) {
+        // These are all blocking calls, so offload the work to another thread
+        TaskPools.ioPool.submit {
+            HomePage(chromeDriver).visit(callUrlInfo.baseUrl)
 
-        val localStorageValues = mutableMapOf(
-            "displayname" to jibriSeleniumOptions.displayName,
-            "email" to jibriSeleniumOptions.email,
-            "callStatsUserName" to "jibri"
-        )
-        xmppCredentials?.let {
-            localStorageValues["xmpp_username_override"] = "${xmppCredentials.username}@${xmppCredentials.domain}"
-            localStorageValues["xmpp_password_override"] = xmppCredentials.password
+            val localStorageValues = mutableMapOf(
+                    "displayname" to jibriSeleniumOptions.displayName,
+                    "email" to jibriSeleniumOptions.email,
+                    "callStatsUserName" to "jibri"
+            )
+            xmppCredentials?.let {
+                localStorageValues["xmpp_username_override"] = "${xmppCredentials.username}@${xmppCredentials.domain}"
+                localStorageValues["xmpp_password_override"] = xmppCredentials.password
+            }
+            setLocalStorageValues(localStorageValues)
+            if (!CallPage(chromeDriver).visit(callUrlInfo.callUrl)) {
+                stateMachine.transition(SeleniumEvent.FailedToJoinCall)
+            } else {
+                startRecurringCallStatusChecks()
+                addParticipantTracker()
+                currCallUrl = callUrlInfo.callUrl
+                stateMachine.transition(SeleniumEvent.CallJoined)
+            }
         }
-        setLocalStorageValues(localStorageValues)
-        if (!CallPage(chromeDriver).visit(callUrlInfo.callUrl)) {
-            return false
-        }
-        startRecurringCallStatusChecks()
-        addParticipantTracker()
-        currCallUrl = callUrlInfo.callUrl
-        return true
     }
 
     fun getParticipants(): List<Map<String, Any>> {
@@ -278,12 +293,12 @@ class JibriSelenium(
     }
 
     private interface CallStatusCheck {
-        fun run(callPage: CallPage): CallStatusCheckResult
+        fun run(callPage: CallPage): SeleniumEvent?
     }
 
     private class EmptyCallStatusCheck : CallStatusCheck {
         private var numTimesEmpty = 0
-        override fun run(callPage: CallPage): CallStatusCheckResult {
+        override fun run(callPage: CallPage): SeleniumEvent? {
             // >1 since the count will include jibri itself
             if (callPage.getNumParticipants() > 1) {
                 numTimesEmpty = 0
@@ -291,15 +306,15 @@ class JibriSelenium(
                 numTimesEmpty++
             }
             if (numTimesEmpty >= 2) {
-                return CallStatusCheckResult.HEALTHY_FINISHED
+                return SeleniumEvent.CallEmpty
             }
-            return CallStatusCheckResult.HEALTHY_ONGOING
+            return null
         }
     }
 
     private class MediaReceivedStatusCheck(private val logger: Logger) : CallStatusCheck {
         private var numTimesNoMedia = 0
-        override fun run(callPage: CallPage): CallStatusCheckResult {
+        override fun run(callPage: CallPage): SeleniumEvent? {
             val bitrates = callPage.getBitrates()
             logger.info("Jibri client receive bitrates: $bitrates")
             val downloadBitrate = bitrates.getOrDefault("download", 0L) as Long
@@ -309,9 +324,9 @@ class JibriSelenium(
                 numTimesNoMedia = 0
             }
             if (numTimesNoMedia >= 2) {
-                return CallStatusCheckResult.UNHEALTHY
+                return SeleniumEvent.NoMediaReceived
             }
-            return CallStatusCheckResult.HEALTHY_ONGOING
+            return null
         }
     }
 }
