+++ src/main/kotlin/org/jitsi/jibri/service/impl/StreamingJibriService.kt
@@ -21,7 +21,7 @@ import net.java.sip.communicator.impl.protocol.jabber.extensions.jibri.JibriIq
 import org.jitsi.jibri.capture.ffmpeg.FfmpegCapturer
 import org.jitsi.jibri.config.XmppCredentials
 import org.jitsi.jibri.selenium.CallParams
-import org.jitsi.jibri.selenium.JibriSelenium2
+import org.jitsi.jibri.selenium.JibriSelenium
 import org.jitsi.jibri.selenium.RECORDING_URL_OPTIONS
 import org.jitsi.jibri.service.JibriService
 import org.jitsi.jibri.service.JibriServiceStateMachine
@@ -73,7 +73,7 @@ class StreamingJibriService(private val streamingParams: StreamingParams) : Jibr
     private val capturer = FfmpegCapturer()
     private val sink: Sink
     private val stateMachine = JibriServiceStateMachine()
-    private val jibriSelenium = JibriSelenium2()
+    private val jibriSelenium = JibriSelenium()
 
     init {
         sink = StreamSink(
@@ -84,9 +84,9 @@ class StreamingJibriService(private val streamingParams: StreamingParams) : Jibr
 
         stateMachine.onStateTransition(this::onServiceStateChange)
 
-        stateMachine.registerSubComponent(JibriSelenium2.COMPONENT_ID)
+        stateMachine.registerSubComponent(JibriSelenium.COMPONENT_ID)
         jibriSelenium.addStatusHandler { state ->
-            stateMachine.transition(state.toJibriServiceEvent(JibriSelenium2.COMPONENT_ID))
+            stateMachine.transition(state.toJibriServiceEvent(JibriSelenium.COMPONENT_ID))
         }
 
         stateMachine.registerSubComponent(FfmpegCapturer.COMPONENT_ID)
